updated new flow admin style

+++ wp-content/plugins/woocommerce-payments/includes/wc-payment-api/class-wc-payments-api-client.php
@@ -43,7 +43,7 @@ class WC_Payments_API_Client {
 	const CAPABILITIES_API             = 'accounts/capabilities';
 	const WOOPAY_ACCOUNTS_API          = 'accounts/platform_checkout';
 	const WOOPAY_COMPATIBILITY_API     = 'woopay/compatibility';
-	const APPLE_PAY_API                = 'apple_pay';
+	const DOMAIN_REGISTRATION_API      = 'payment_method_domains';
 	const CHARGES_API                  = 'charges';
 	const CONN_TOKENS_API              = 'terminal/connection_tokens';
 	const TERMINAL_LOCATIONS_API       = 'terminal/locations';
@@ -79,6 +79,7 @@ class WC_Payments_API_Client {
 	const FRAUD_OUTCOMES_API           = 'fraud_outcomes';
 	const FRAUD_RULESET_API            = 'fraud_ruleset';
 	const COMPATIBILITY_API            = 'compatibility';
+	const REPORTING_API                = 'reporting/payment_activity';
 
 	/**
 	 * Common keys in API requests/responses that we might want to redact.
@@ -230,8 +231,17 @@ class WC_Payments_API_Client {
 	 * @param string $intent_id intent id.
 	 *
 	 * @return WC_Payments_API_Payment_Intention intention object.
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function get_intent( $intent_id ) {
+		if ( ! preg_match( '/^\w+$/', $intent_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$intent = $this->request( [], self::INTENTIONS_API . '/' . $intent_id, self::GET );
 
 		return $this->deserialize_payment_intention_object_from_array( $intent );
@@ -447,8 +457,17 @@ class WC_Payments_API_Client {
 	 *
 	 * @param string $transaction_id id of requested transaction.
 	 * @return array transaction object.
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function get_transaction( $transaction_id ) {
+		if ( ! preg_match( '/^\w+$/', $transaction_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$transaction = $this->request( [], self::TRANSACTIONS_API . '/' . $transaction_id, self::GET );
 
 		if ( is_wp_error( $transaction ) ) {
@@ -517,8 +536,17 @@ class WC_Payments_API_Client {
 	 *
 	 * @param string $dispute_id id of requested dispute.
 	 * @return array dispute object.
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function get_dispute( $dispute_id ) {
+		if ( ! preg_match( '/^\w+$/', $dispute_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$dispute = $this->request( [], self::DISPUTES_API . '/' . $dispute_id, self::GET );
 
 		if ( is_wp_error( $dispute ) ) {
@@ -538,8 +566,17 @@ class WC_Payments_API_Client {
 	 * @param array  $metadata   metadata associated with this dispute.
 	 *
 	 * @return array dispute object.
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function update_dispute( $dispute_id, $evidence, $submit, $metadata ) {
+		if ( ! preg_match( '/^\w+$/', $dispute_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$request = [
 			'evidence' => $evidence,
 			'submit'   => $submit,
@@ -564,8 +601,18 @@ class WC_Payments_API_Client {
 	 *
 	 * @param string $dispute_id id of dispute to close.
 	 * @return array dispute object.
+	 *
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function close_dispute( $dispute_id ) {
+		if ( ! preg_match( '/^\w+$/', $dispute_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$dispute = $this->request( [], self::DISPUTES_API . '/' . $dispute_id . '/close', self::POST );
 		// Invalidate the dispute caches.
 		\WC_Payments::get_database_cache()->delete( Database_Cache::DISPUTE_STATUS_COUNTS_KEY );
@@ -685,6 +732,14 @@ class WC_Payments_API_Client {
 	 */
 	public function get_file_contents( string $file_id, bool $as_account = true ): array {
 		try {
+			if ( ! preg_match( '/^\w+$/', $file_id ) ) {
+				throw new API_Exception(
+					__( 'Route param validation failed.', 'woocommerce-payments' ),
+					'wcpay_route_validation_failure',
+					400
+				);
+			}
+
 			return $this->request( [ 'as_account' => $as_account ], self::FILES_API . '/' . $file_id . '/contents', self::GET );
 		} catch ( API_Exception $e ) {
 			Logger::error( 'Error retrieving file contents for ' . $file_id . '. ' . $e->getMessage() );
@@ -702,6 +757,13 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception
 	 */
 	public function get_file( string $file_id, bool $as_account = true ): array {
+		if ( ! preg_match( '/^\w+$/', $file_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
 		return $this->request( [ 'as_account' => $as_account ], self::FILES_API . '/' . $file_id, self::GET );
 	}
 
@@ -725,8 +787,17 @@ class WC_Payments_API_Client {
 	 * @return array
 	 *
 	 * @throws Exception - Exception thrown on request failure.
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function get_timeline( $id ) {
+		if ( ! preg_match( '/^\w+$/', $id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$timeline = $this->request( [], self::TIMELINE_API . '/' . $id, self::GET );
 
 		$has_fraud_outcome_event = false;
@@ -954,36 +1025,6 @@ class WC_Payments_API_Client {
 	}
 
 	/**
-	 * Get the required verification information, needed for our KYC onboarding flow.
-	 *
-	 * @param string      $country_code The country code.
-	 * @param string      $type         The business type.
-	 * @param string|null $structure    The business structure (optional).
-	 *
-	 * @return array An array containing the required verification information.
-	 *
-	 * @throws API_Exception Exception thrown on request failure.
-	 */
-	public function get_onboarding_required_verification_information( string $country_code, string $type, $structure = null ) {
-		$params = [
-			'country' => $country_code,
-			'type'    => $type,
-		];
-
-		if ( ! is_null( $structure ) ) {
-			$params = array_merge( $params, [ 'structure' => $structure ] );
-		}
-
-		return $this->request(
-			$params,
-			self::ONBOARDING_API . '/required_verification_information',
-			self::GET,
-			true,
-			true
-		);
-	}
-
-	/**
 	 * Get a link's details from the server.
 	 *
 	 * @param array $args The arguments to be sent with the link request.
@@ -1045,6 +1086,14 @@ class WC_Payments_API_Client {
 			);
 		}
 
+		if ( ! preg_match( '/^\w+$/', $customer_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$this->request(
 			$customer_data,
 			self::CUSTOMERS_API . '/' . $customer_id,
@@ -1088,6 +1137,14 @@ class WC_Payments_API_Client {
 			);
 		}
 
+		if ( ! preg_match( '/^\w+$/', $product_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$product_data,
 			self::PRODUCTS_API . '/' . $product_id,
@@ -1112,6 +1169,14 @@ class WC_Payments_API_Client {
 			);
 		}
 
+		if ( ! preg_match( '/^\w+$/', $price_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$this->request(
 			$price_data,
 			self::PRICES_API . '/' . $price_id,
@@ -1129,6 +1194,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If fetching the invoice fails.
 	 */
 	public function get_invoice( string $invoice_id ) {
+		if ( ! preg_match( '/^\w+$/', $invoice_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::INVOICES_API . '/' . $invoice_id,
@@ -1148,6 +1221,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception Error charging the invoice.
 	 */
 	public function charge_invoice( string $invoice_id, array $data = [] ) {
+		if ( ! preg_match( '/^\w+$/', $invoice_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::INVOICES_API . '/' . $invoice_id . '/pay',
@@ -1165,6 +1246,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception Error updating the invoice.
 	 */
 	public function update_invoice( string $invoice_id, array $data = [] ) {
+		if ( ! preg_match( '/^\w+$/', $invoice_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::INVOICES_API . '/' . $invoice_id,
@@ -1182,6 +1271,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception
 	 */
 	public function update_charge( string $charge_id, array $data = [] ) {
+		if ( ! preg_match( '/^\w+$/', $charge_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::CHARGES_API . '/' . $charge_id,
@@ -1198,6 +1295,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception
 	 */
 	public function get_charge( string $charge_id ) {
+		if ( ! preg_match( '/^\w+$/', $charge_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::CHARGES_API . '/' . $charge_id,
@@ -1215,6 +1320,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception
 	 */
 	public function update_transaction( string $transaction_id, array $data = [] ) {
+		if ( ! preg_match( '/^\w+$/', $transaction_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::TRANSACTIONS_API . '/' . $transaction_id,
@@ -1232,6 +1345,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If fetching the subscription fails.
 	 */
 	public function get_subscription( string $wcpay_subscription_id ) {
+		if ( ! preg_match( '/^\w+$/', $wcpay_subscription_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::SUBSCRIPTIONS_API . '/' . $wcpay_subscription_id,
@@ -1267,6 +1388,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If updating the WCPay subscription fails.
 	 */
 	public function update_subscription( $wcpay_subscription_id, $data ) {
+		if ( ! preg_match( '/^\w+$/', $wcpay_subscription_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::SUBSCRIPTIONS_API . '/' . $wcpay_subscription_id,
@@ -1284,6 +1413,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If canceling the subscription fails.
 	 */
 	public function cancel_subscription( string $wcpay_subscription_id ) {
+		if ( ! preg_match( '/^\w+$/', $wcpay_subscription_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::SUBSCRIPTIONS_API . '/' . $wcpay_subscription_id,
@@ -1302,6 +1439,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If updating the WCPay subscription item fails.
 	 */
 	public function update_subscription_item( $wcpay_subscription_item_id, $data ) {
+		if ( ! preg_match( '/^\w+$/', $wcpay_subscription_item_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::SUBSCRIPTION_ITEMS_API . '/' . $wcpay_subscription_item_id,
@@ -1319,6 +1464,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If payment method does not exist.
 	 */
 	public function get_payment_method( $payment_method_id ) {
+		if ( ! preg_match( '/^\w+$/', $payment_method_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::PAYMENT_METHODS_API . '/' . $payment_method_id,
@@ -1337,6 +1490,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If payment method update fails.
 	 */
 	public function update_payment_method( $payment_method_id, $payment_method_data = [] ) {
+		if ( ! preg_match( '/^\w+$/', $payment_method_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$payment_method_data,
 			self::PAYMENT_METHODS_API . '/' . $payment_method_id,
@@ -1356,6 +1517,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If an error occurs.
 	 */
 	public function get_payment_methods( $customer_id, $type, $limit = 100 ) {
+		if ( ! preg_match( '/^\w+$/', $customer_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[
 				'customer' => $customer_id,
@@ -1377,6 +1546,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If detachment fails.
 	 */
 	public function detach_payment_method( $payment_method_id ) {
+		if ( ! preg_match( '/^\w+$/', $payment_method_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::PAYMENT_METHODS_API . '/' . $payment_method_id . '/detach',
@@ -1426,22 +1603,28 @@ class WC_Payments_API_Client {
 		);
 	}
 
+
 	/**
-	 * Registers a new domain with Apple Pay.
+	 * Registers a Payment Method Domain.
 	 *
-	 * @param string $domain_name Domain name which to register for Apple Pay.
+	 * @param string $domain_name Domain name which to register for the account.
 	 *
-	 * @return array An array containing an id in case it has succeeded, or an error message in case it has failed.
+	 * @return array An array containing an id and the bool property 'enabled' indicating
+	 * whether the domain is enabled for the account. Each Payment Method
+	 * (apple_pay, google_pay, link, paypal) in the array have a 'status'
+	 * property with the possible values 'active' and 'inactive'.
 	 *
 	 * @throws API_Exception If an error occurs.
 	 */
-	public function register_domain_with_apple( $domain_name ) {
+	public function register_domain( $domain_name ) {
 		return $this->request(
 			[
-				'test_mode'   => false, // Force live mode - Domain registration doesn't work in test mode.
 				'domain_name' => $domain_name,
+				// The value needs to be a string.
+				// If it's a boolean, it gets serialized as an integer (1), causing an invalid request error.
+				'enabled'     => 'true',
 			],
-			self::APPLE_PAY_API . '/domains',
+			self::DOMAIN_REGISTRATION_API,
 			self::POST
 		);
 	}
@@ -1542,6 +1725,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If an error occurs.
 	 */
 	public function update_terminal_location( $location_id, $display_name, $address ) {
+		if ( ! preg_match( '/^\w+$/', $location_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		// Any parameters not provided will be left unchanged so pass only supplied values.
 		$update_request_body = array_merge(
 			( isset( $address ) ? [ 'address' => $address ] : [] ),
@@ -1566,6 +1757,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If the location id is invalid or downstream call fails.
 	 */
 	public function delete_terminal_location( $location_id ) {
+		if ( ! preg_match( '/^\w+$/', $location_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request( [], self::TERMINAL_LOCATIONS_API . '/' . $location_id, self::DELETE );
 	}
 
@@ -1602,6 +1801,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception - If not connected or request failed.
 	 */
 	public function get_document( $document_id ) {
+		if ( ! preg_match( '/^[\w-]+$/', $document_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request( [], self::DOCUMENTS_API . '/' . $document_id, self::GET, true, false, true );
 	}
 
@@ -1680,6 +1887,14 @@ class WC_Payments_API_Client {
 	 * @return array The response object.
 	 */
 	public function get_latest_fraud_outcome( $id ) {
+		if ( ! preg_match( '/^\w+$/', $id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$response = $this->request(
 			[],
 			self::FRAUD_OUTCOMES_API . '/order_id/' . $id,
@@ -1859,6 +2074,9 @@ class WC_Payments_API_Client {
 		while ( true ) {
 			$response_code  = null;
 			$last_exception = null;
+
+			// The header intention is to give us insights into request latency between store and backend.
+			$headers['X-Request-Initiated'] = microtime( true );
 
 			try {
 				$response = $this->http_client->remote_request(
@@ -2458,6 +2676,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If an error occurs.
 	 */
 	public function get_currency_minimum_recurring_amount( $currency ) {
+		if ( ! preg_match( '/^\w+$/', $currency ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return (int) $this->request(
 			[],
 			self::MINIMUM_RECURRING_AMOUNT_API . '/' . $currency,
@@ -2496,8 +2722,17 @@ class WC_Payments_API_Client {
 	 *
 	 * @param string $payment_intent_id id of requested transaction.
 	 * @return array authorization object.
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function get_authorization( string $payment_intent_id ) {
+		if ( ! preg_match( '/^\w+$/', $payment_intent_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request( [], self::AUTHORIZATIONS_API . '/' . $payment_intent_id, self::GET );
 	}
